+++ EZEats/.idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/EasyEats.iml" filepath="$PROJECT_DIR$/EasyEats.iml" />
+      <module fileurl="file://$PROJECT_DIR$/EZEats.iml" filepath="$PROJECT_DIR$/EZEats.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
     </modules>
   </component>
