+++ EZEats/app/build.gradle
@@ -5,7 +5,7 @@ android {
     buildToolsVersion "23.0.2"
 
     defaultConfig {
-        applicationId "edu.ucsd.cse110_team11.easyeats"
+        applicationId "edu.ucsd.cse110_team11.ezeats"
         minSdkVersion 16
         targetSdkVersion 23
         versionCode 1
