+++ EZEats/app/src/androidTest/java/edu/ucsd/cse110_team11/ezeats/ApplicationTest.java
@@ -1,4 +1,4 @@
-package edu.ucsd.cse110_team11.easyeats;
+package edu.ucsd.cse110_team11.ezeats;
 
 import android.app.Application;
 import android.test.ApplicationTestCase;
