--- EasyEats/.idea/.name
@@ -1,1 +1,0 @@
-Easy Eats
